destroy

+++ resources/views/tasks/show.blade.php
@@ -19,10 +19,11 @@
             </tr>
         </table>
 
+@if (Auth::check())
         {!! link_to_route('tasks.edit', 'このタスク編集', ['id' => $task->id], ['class' => 'btn btn-default']) !!}
-        
+       
         {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' =>'delete']) !!}
         {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
         {!! Form::close() !!}
-
+    @endif
 @endsection
